// src/shared/api/generated/model/updateProfileInputDto.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * SnapFlow API
 * REST API для SnapFlow
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateProfileInputDto {
  /**
   * @minLength 6
   * @maxLength 30
   * @pattern ^[a-zA-Z0-9_-]*$
   */
  username: string;
  /**
   * @minLength 1
   * @maxLength 50
   * @pattern ^[А-ЯЁа-яёA-Za-z]+(?:\s*[А-ЯЁа-яёA-Za-z]+)*$
   */
  firstName: string;
  /**
   * @minLength 1
   * @maxLength 50
   * @pattern ^[А-ЯЁа-яёA-Za-z]+(?:\s*[А-ЯЁа-яёA-Za-z]+)*$
   */
  lastName: string;
  /**
   * ISO date string
   * @nullable
   */
  dateOfBirth?: string | null;
  /**
   * @minLength 1
   * @maxLength 50
   * @nullable
   */
  country?: string | null;
  /**
   * @minLength 1
   * @maxLength 50
   * @nullable
   */
  city?: string | null;
  /**
   * @minLength 0
   * @maxLength 200
   * @nullable
   */
  aboutMe?: string | null;
}